requirements.txt
pyyaml
python-dotenv
google-generativeai
tabulate
StrEnum
